PCI: Configure RC MPS to 128 for Realtek 8168 attach
Change is to configure QC RC MPS 128 bytes as 8168 device only supports
128 bytes. Without this quirk, since RC is using pcie_bus_perf flag
MPS is configured to 256 bytes, when EP MRRS is changed to 4096 bytes
RC is responding with packets of 256 bytes causing functionality break.

Since pcie_bus_perf is used for many products, this change provies a
provision to configure RC to 128 bytes for realtek 8168 attach alone.

Change-Id: I4a82293e4f71bf188d4f0f2be8fa194c5ab007c2
Signed-off-by: Subramanian Ananthanarayanan <quic_skananth@quicinc.com>