firmware: qcom_scm: handle echo b > /proc/sysrq-trigger
Introduce restart handler to handle "echo b > /proc/sysrq-trigger".
One special thing with 'b' kind of reboot is, it does not call reboot
notifiers and call restart handler right a way and that does not seems
to work for blair SoC and it goes to dump mode.

Keep this restart handler priority to 201 greater than 200 which is the
priority for msm-poweroff to make reboot work cleaner.

Change-Id: I42f8c0fde29af402096d760c37ec2d8b3a85439a
Signed-off-by: Mukesh Ojha <quic_mojha@quicinc.com>
Signed-off-by: Sayan Dey <quic_sayand@quicinc.com>